requirements.txt
pytorch-lightning==1.4.0
torchmetrics==0.5
omegaconf
einops
tqdm
albumentations
pandas
natsort